requirements.txt
asgiref==3.3.4
boto3==1.17.53
botocore==1.20.53
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
jmespath==0.10.0
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.7
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4

tensorflow~=2.7.0
numpy~=1.21.5
torchvision~=0.11.2